Moving pre_processing steps logic to OmenMarket | moving graph queries to separate file
