+++ prediction_market_agent_tooling/markets/omen/omen.py
@@ -8,7 +8,8 @@ from web3 import Web3
 from web3.constants import HASH_ZERO
 
 from prediction_market_agent_tooling.config import APIKeys
-from prediction_market_agent_tooling.deploy.agent import MAX_AVAILABLE_MARKETS
+
+# from prediction_market_agent_tooling.deploy.agent import MAX_AVAILABLE_MARKETS
 from prediction_market_agent_tooling.gtypes import (
     ChecksumAddress,
     HexAddress,
@@ -24,14 +25,12 @@ from prediction_market_agent_tooling.markets.agent_market import (
     SortBy,
 )
 from prediction_market_agent_tooling.markets.data_models import BetAmount, Currency
-from prediction_market_agent_tooling.markets.markets import MarketType
 from prediction_market_agent_tooling.markets.omen.data_models import (
     OMEN_FALSE_OUTCOME,
     OMEN_TRUE_OUTCOME,
     Condition,
     OmenBet,
     OmenMarket,
-    OmenUserPosition,
 )
 from prediction_market_agent_tooling.markets.omen.omen_contracts import (
     Arbitrator,
@@ -41,30 +40,18 @@ from prediction_market_agent_tooling.markets.omen.omen_contracts import (
     OmenFixedProductMarketMakerFactoryContract,
     OmenOracleContract,
     OmenRealitioContract,
+    OMEN_DEFAULT_MARKET_FEE,
 )
 from prediction_market_agent_tooling.markets.omen.omen_graph_queries import (
-    get_user_positions,
+    get_omen_markets,
+    get_omen_bets,
 )
-from prediction_market_agent_tooling.tools.utils import utcnow
+
 from prediction_market_agent_tooling.tools.web3_utils import (
     add_fraction,
     private_key_to_public_key,
     remove_fraction,
     xdai_to_wei,
-)
-
-"""
-Python API for Omen prediction market.
-
-Their API is available as graph on https://thegraph.com/explorer/subgraphs/9V1aHHwkK4uPWgBH6ZLzwqFEkoHTHPS7XHKyjZWe8tEf?view=Overview&chain=mainnet,
-but to not use our own credits, seems we can use their api deployment directly: https://api.thegraph.com/subgraphs/name/protofire/omen-xdai/graphql (link to the online playground)
-"""
-
-OMEN_QUERY_BATCH_SIZE = 1000
-OMEN_DEFAULT_MARKET_FEE = 0.02  # 2% fee from the buying shares amount.
-OMEN_TRADES_SUBGRAPH = "https://api.thegraph.com/subgraphs/name/protofire/omen-xdai"
-CONDITIONAL_TOKENS_SUBGRAPH = (
-    "https://api.thegraph.com/subgraphs/name/gnosis/conditional-tokens-gc"
 )
 
 
@@ -78,7 +65,7 @@ class OmenAgentMarket(AgentMarket):
     market_maker_contract_address_checksummed: ChecksumAddress
     condition: Condition
 
-    INVALID_MARKET_ANSWER = (
+    INVALID_MARKET_ANSWER: HexStr = (
         0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF
     )
 
@@ -120,7 +107,7 @@ class OmenAgentMarket(AgentMarket):
 
             # Like Olas, we assert correctness by matching index OR invalid market answer
             if bet.outcomeIndex == int(
-                bet.fpmm.question.currentAnswer
+                bet.fpmm.question.currentAnswer, 16
             ) or bet.outcomeIndex == int(self.INVALID_MARKET_ANSWER):
                 return True
 
@@ -132,7 +119,7 @@ class OmenAgentMarket(AgentMarket):
         Since we currently use stateless functions to redeem positions, it's not possible to query state. Hence we proceed by not tracking the positions already redeemed.
         Note that this has no major consequences from a gas perspective, it only incurs extra subgraph queries and RPC reads, no writes hence no gas costs.
         """
-        pass
+        return False
 
     def redeem_positions(self) -> None:
         keys = APIKeys()
@@ -143,20 +130,19 @@ class OmenAgentMarket(AgentMarket):
             return None
 
         position_already_redeemed = self.check_if_position_was_already_redeemed()
-        if not position_already_redeemed:
+        if position_already_redeemed:
             print(f"Position on market {self.id} was already redeemed.")
             return None
 
         return omen_redeem_full_position_tx(
-            market=self,
-            from_private_key=keys.bet_from_private_key,
+            market=self, from_private_key=keys.bet_from_private_key
         )
 
     def before_process_bets(self) -> None:
         # We can only redeem positions from resolved markets.
         resolved_markets = self.get_binary_markets(
             filter_by=FilterBy.RESOLVED,
-            limit=MAX_AVAILABLE_MARKETS,
+            limit=20,  # MAX_AVAILABLE_MARKETS,
             sort_by=SortBy.CLOSING_SOONEST,
         )
         for market in resolved_markets:
@@ -208,143 +194,6 @@ class OmenAgentMarket(AgentMarket):
         )
 
 
-_QUERY_GET_SINGLE_FIXED_PRODUCT_MARKET_MAKER = """
-query getFixedProductMarketMaker($id: String!) {
-    fixedProductMarketMaker(
-        id: $id
-    ) {
-        id
-        title
-        category
-        creationTimestamp
-        collateralVolume
-        usdVolume
-        collateralToken
-        outcomes
-        outcomeTokenAmounts
-        outcomeTokenMarginalPrices
-        fee
-        condition {
-            id
-            outcomeSlotCount
-        }
-        answerFinalizedTimestamp
-        resolutionTimestamp
-        currentAnswer
-    }
-}
-"""
-
-
-def construct_query_get_fixed_product_markets_makers(
-    include_creator: bool, filter_by: FilterBy
-) -> str:
-    query = """
-        query getFixedProductMarketMakers(
-            $first: Int!,
-            $outcomes: [String!],
-            $orderBy: String!,
-            $orderDirection: String!,
-            $creationTimestamp_gt: Int!,
-            $creator: Bytes = null,
-        ) {
-            fixedProductMarketMakers(
-                where: {
-                    isPendingArbitration: false,
-                    outcomes: $outcomes
-                    creationTimestamp_gt: $creationTimestamp_gt
-                    creator: $creator,
-                    answerFinalizedTimestamp: null
-                    resolutionTimestamp_not: null
-                },
-                orderBy: creationTimestamp,
-                orderDirection: desc,
-                first: $first
-            ) {
-                id
-                title
-                collateralVolume
-                usdVolume
-                collateralToken
-                outcomes
-                outcomeTokenAmounts
-                outcomeTokenMarginalPrices
-                fee
-                answerFinalizedTimestamp
-                resolutionTimestamp
-                currentAnswer
-                creationTimestamp
-                category
-                condition {
-                    id
-                    outcomeSlotCount
-                }
-            }
-        }
-    """
-
-    if filter_by == FilterBy.OPEN:
-        query = query.replace("resolutionTimestamp_not: null", "")
-    elif filter_by == FilterBy.RESOLVED:
-        query = query.replace("answerFinalizedTimestamp: null", "")
-    elif filter_by == FilterBy.NONE:
-        query = query.replace("answerFinalizedTimestamp: null", "")
-        query = query.replace("resolutionTimestamp_not: null", "")
-    else:
-        raise ValueError(f"Unknown filter_by: {filter_by}")
-
-    if not include_creator:
-        # If we aren't filtering by query, we need to remove it from where, otherwise "creator: null" will return 0 results.
-        query = query.replace("creator: $creator,", "")
-
-    return query
-
-
-def ordering_from_sort_by(sort_by: SortBy) -> tuple[str, str]:
-    """
-    Returns 'orderBy' and 'orderDirection' strings for the given SortBy.
-    """
-    if sort_by == SortBy.CLOSING_SOONEST:
-        return "creationTimestamp", "desc"  # TODO make more accurate
-    elif sort_by == SortBy.NEWEST:
-        return "creationTimestamp", "desc"
-    else:
-        raise ValueError(f"Unknown sort_by: {sort_by}")
-
-
-def get_omen_markets(
-    first: int,
-    outcomes: list[str],
-    sort_by: SortBy,
-    filter_by: FilterBy,
-    created_after: t.Optional[datetime] = None,
-    creator: t.Optional[HexAddress] = None,
-) -> list[OmenMarket]:
-    order_by, order_direction = ordering_from_sort_by(sort_by)
-    markets = requests.post(
-        OMEN_TRADES_SUBGRAPH,
-        json={
-            "query": construct_query_get_fixed_product_markets_makers(
-                include_creator=creator is not None,
-                filter_by=filter_by,
-            ),
-            "variables": {
-                "first": first,
-                "outcomes": outcomes,
-                "orderBy": order_by,
-                "orderDirection": order_direction,
-                "creationTimestamp_gt": (
-                    to_int_timestamp(created_after) if created_after else 0
-                ),
-                "creator": creator,
-            },
-        },
-        headers={"Content-Type": "application/json"},
-    ).json()
-    markets = markets["data"]["fixedProductMarketMakers"]
-    return [OmenMarket.model_validate(market) for market in markets]
-
-
 def get_omen_binary_markets(
     limit: int,
     sort_by: SortBy,
@@ -366,20 +215,6 @@ def pick_binary_market(
     sort_by: SortBy = SortBy.CLOSING_SOONEST, filter_by: FilterBy = FilterBy.OPEN
 ) -> OmenMarket:
     return get_omen_binary_markets(limit=1, sort_by=sort_by, filter_by=filter_by)[0]
-
-
-def get_market(market_id: str) -> OmenMarket:
-    market = requests.post(
-        OMEN_TRADES_SUBGRAPH,
-        json={
-            "query": _QUERY_GET_SINGLE_FIXED_PRODUCT_MARKET_MAKER,
-            "variables": {
-                "id": market_id,
-            },
-        },
-        headers={"Content-Type": "application/json"},
-    ).json()["data"]["fixedProductMarketMaker"]
-    return OmenMarket.model_validate(market)
 
 
 def omen_buy_outcome_tx(
@@ -517,115 +352,6 @@ def binary_omen_sell_outcome_tx(
         outcome=OMEN_TRUE_OUTCOME if binary_outcome else OMEN_FALSE_OUTCOME,
         auto_withdraw=auto_withdraw,
     )
-
-
-# Order by id, so we can use id_gt for pagination.
-_QUERY_GET_FIXED_PRODUCT_MARKETS_MAKER_TRADES = """
-query getFixedProductMarketMakerTrades(
-    $id_gt: String!,
-    $creator: String!,
-    $creationTimestamp_gte: Int!,
-    $creationTimestamp_lte: Int!,
-    $first: Int!,
-) {
-    fpmmTrades(
-        where: {
-            type: Buy,
-            creator: $creator,
-            creationTimestamp_gte: $creationTimestamp_gte,
-            creationTimestamp_lte: $creationTimestamp_lte,
-            id_gt: $id_gt,
-        }
-        first: $first
-        orderBy: id
-        orderDirection: asc
-    ) {
-        id
-        title
-        collateralToken
-        outcomeTokenMarginalPrice
-        oldOutcomeTokenMarginalPrice
-        type
-        creator {
-            id
-        }
-        creationTimestamp
-        collateralAmount
-        collateralAmountUSD
-        feeAmount
-        outcomeIndex
-        outcomeTokensTraded
-        transactionHash
-        fpmm {
-            id
-            outcomes
-            title
-            answerFinalizedTimestamp
-            resolutionTimestamp
-            currentAnswer
-            isPendingArbitration
-            arbitrationOccurred
-            openingTimestamp
-            condition {
-                id
-                outcomeSlotCount
-            }
-            collateralVolume
-            usdVolume
-            collateralToken
-            outcomeTokenAmounts
-            outcomeTokenMarginalPrices
-            fee
-            category
-        }
-    }
-}
-"""
-
-
-def to_int_timestamp(dt: datetime) -> int:
-    return int(dt.timestamp())
-
-
-def get_omen_bets(
-    better_address: ChecksumAddress,
-    start_time: datetime,
-    end_time: t.Optional[datetime],
-) -> list[OmenBet]:
-    if not end_time:
-        end_time = utcnow()
-
-    # Initialize id_gt for the first batch of bets to zero
-    id_gt: str = "0"
-    all_bets: list[OmenBet] = []
-    while True:
-        query = _QUERY_GET_FIXED_PRODUCT_MARKETS_MAKER_TRADES
-        bets = requests.post(
-            OMEN_TRADES_SUBGRAPH,
-            json={
-                "query": query,
-                "variables": {
-                    "creator": better_address.lower(),
-                    "creationTimestamp_gte": to_int_timestamp(start_time),
-                    "creationTimestamp_lte": to_int_timestamp(end_time),
-                    "id_gt": id_gt,
-                    "first": OMEN_QUERY_BATCH_SIZE,
-                },
-            },
-            headers={"Content-Type": "application/json"},
-        ).json()
-
-        bets = bets.get("data", {}).get("fpmmTrades", [])
-
-        if not bets:
-            break
-
-        # Increment id_gt for the next batch of bets
-        id_gt = bets[-1]["id"]
-
-        all_bets.extend(OmenBet.model_validate(bet) for bet in bets)
-
-    return all_bets
 
 
 def get_resolved_omen_bets(
