+++ scripts/create_market_omen.py
@@ -8,8 +8,10 @@ from prediction_market_agent_tooling.markets.omen.data_models import (
     OMEN_TRUE_OUTCOME,
 )
 from prediction_market_agent_tooling.markets.omen.omen import (
-    OMEN_DEFAULT_MARKET_FEE,
     omen_create_market_tx,
+)
+from prediction_market_agent_tooling.markets.omen.omen_contracts import (
+    OMEN_DEFAULT_MARKET_FEE,
 )
 from prediction_market_agent_tooling.tools.web3_utils import verify_address
