+++ tests/markets/test_omen.py
@@ -13,7 +13,6 @@ from prediction_market_agent_tooling.markets.omen.omen import (
     OmenAgentMarket,
     binary_omen_buy_outcome_tx,
     binary_omen_sell_outcome_tx,
-    get_market,
     get_omen_bets,
     get_omen_binary_markets,
     get_resolved_omen_bets,
@@ -23,6 +22,7 @@ from prediction_market_agent_tooling.markets.omen.omen import (
     omen_remove_fund_market_tx,
     pick_binary_market,
 )
+from prediction_market_agent_tooling.markets.omen.omen_graph_queries import get_market
 from prediction_market_agent_tooling.tools.contract import wait_until_nonce_changed
 from prediction_market_agent_tooling.tools.utils import check_not_none
 from prediction_market_agent_tooling.tools.web3_utils import xdai_to_wei
